requirements.txt
streamlit
confluent-kafka
pandas
river
altair
python-dotenv